+++ lib/src/internals/refresh_physics.dart
@@ -110,8 +110,8 @@ class RefreshPhysics extends ScrollPhysics {
   double applyPhysicsToUserOffset(ScrollMetrics position, double offset) {
     // TODO: implement applyPhysicsToUserOffset
     viewportRender ??=
-        findViewport(controller!.position?.context.storageContext);
-    if (controller!.headerMode!.value == RefreshStatus.twoLeveling) {
+        findViewport(controller?.position?.context.storageContext);
+    if (controller?.headerStatus == RefreshStatus.twoLeveling) {
       if (offset > 0.0) {
         return parent!.applyPhysicsToUserOffset(position, offset);
       }
@@ -123,12 +123,12 @@ class RefreshPhysics extends ScrollPhysics {
       }
     }
     if (position.outOfRange ||
-        controller!.headerMode!.value == RefreshStatus.twoLeveling) {
+        controller?.headerStatus == RefreshStatus.twoLeveling) {
       final double overscrollPastStart =
           math.max(position.minScrollExtent - position.pixels, 0.0);
       final double overscrollPastEnd = math.max(
           position.pixels -
-              (controller!.headerMode!.value == RefreshStatus.twoLeveling
+              (controller?.headerStatus == RefreshStatus.twoLeveling
                   ? 0.0
                   : position.maxScrollExtent),
           0.0);
@@ -170,7 +170,7 @@ class RefreshPhysics extends ScrollPhysics {
   double applyBoundaryConditions(ScrollMetrics position, double value) {
     final ScrollPosition scrollPosition = position as ScrollPosition;
     viewportRender ??=
-        findViewport(controller!.position?.context.storageContext);
+        findViewport(controller?.position?.context.storageContext);
     bool notFull = position.minScrollExtent == position.maxScrollExtent;
     final bool enablePullDown = viewportRender == null
         ? false
@@ -178,7 +178,7 @@ class RefreshPhysics extends ScrollPhysics {
     final bool enablePullUp = viewportRender == null
         ? false
         : viewportRender!.lastChild is RenderSliverLoading;
-    if (controller!.headerMode!.value == RefreshStatus.twoLeveling) {
+    if (controller?.headerStatus == RefreshStatus.twoLeveling) {
       if (position.pixels - value > 0.0) {
         return parent!.applyBoundaryConditions(position, value);
       }
@@ -200,17 +200,15 @@ class RefreshPhysics extends ScrollPhysics {
     if (enablePullUp) {
       final RenderSliverLoading? sliverFooter =
           viewportRender!.lastChild as RenderSliverLoading?;
+      final storageContext = controller?.position?.context.storageContext;
+      final configuration = storageContext == null
+          ? null
+          : RefreshConfiguration.of(storageContext);
       bottomExtra = (!notFull && sliverFooter!.geometry!.scrollExtent != 0) ||
               (notFull &&
-                  controller!.footerStatus == LoadStatus.noMore &&
-                  !RefreshConfiguration.of(
-                          controller!.position!.context.storageContext)!
-                      .enableLoadingWhenNoData) ||
-              (notFull &&
-                  (RefreshConfiguration.of(
-                              controller!.position!.context.storageContext)
-                          ?.hideFooterWhenNotFull ??
-                      false))
+                  controller?.footerStatus == LoadStatus.noMore &&
+                  !(configuration?.enableLoadingWhenNoData ?? true)) ||
+              (notFull && (configuration?.hideFooterWhenNotFull ?? false))
           ? 0.0
           : sliverFooter!.layoutExtent;
     }
@@ -265,7 +263,7 @@ class RefreshPhysics extends ScrollPhysics {
       ScrollMetrics position, double velocity) {
     // TODO: implement createBallisticSimulation
     viewportRender ??=
-        findViewport(controller!.position?.context.storageContext);
+        findViewport(controller?.position?.context.storageContext);
 
     final bool enablePullDown = viewportRender == null
         ? false
@@ -273,7 +271,7 @@ class RefreshPhysics extends ScrollPhysics {
     final bool enablePullUp = viewportRender == null
         ? false
         : viewportRender!.lastChild is RenderSliverLoading;
-    if (controller!.headerMode!.value == RefreshStatus.twoLeveling) {
+    if (controller?.headerStatus == RefreshStatus.twoLeveling) {
       if (velocity < 0.0) {
         return parent!.createBallisticSimulation(position, velocity);
       }
@@ -284,7 +282,7 @@ class RefreshPhysics extends ScrollPhysics {
       }
     }
     if ((position.pixels > 0 &&
-            controller!.headerMode!.value == RefreshStatus.twoLeveling) ||
+            controller?.headerStatus == RefreshStatus.twoLeveling) ||
         position.outOfRange) {
       return BouncingScrollSimulation(
         spring: springDescription ?? spring,
@@ -293,10 +291,9 @@ class RefreshPhysics extends ScrollPhysics {
         velocity: velocity * 0.91,
         // TODO(abarth): We should move this constant closer to the drag end.
         leadingExtent: position.minScrollExtent,
-        trailingExtent:
-            controller!.headerMode!.value == RefreshStatus.twoLeveling
-                ? 0.0
-                : position.maxScrollExtent,
+        trailingExtent: controller?.headerStatus == RefreshStatus.twoLeveling
+            ? 0.0
+            : position.maxScrollExtent,
         tolerance: toleranceFor(position),
       );
     }